+++ v1_Maintenance/bin/icmp_mon.c
@@ -32,10 +32,13 @@ This file is part of MADCAT, the Mass Attack Detection Acceptance Tool.
 */
 
 //Header includes, defintions and globals
+#include "madcat.common.h"
+#include "madcat.helper.h"
 #include "icmp_mon.h"
 #include "icmp_mon.helper.h"
 #include "icmp_mon.parser.h"
 #include "icmp_mon.worker.h"
+#include "udp_ip_port_mon.icmp_mon.helper.h"
 
 //Main
 int main(int argc, char *argv[])
