+++ v1_Maintenance/bin/udp_ip_port_mon.c
@@ -35,10 +35,13 @@ This file is part of MADCAT, the Mass Attack Detection Acceptance Tool.
 */
 
 //Header includes, defintions and globals
+#include "madcat.helper.h"
+#include "madcat.common.h"
 #include "udp_ip_port_mon.h"
 #include "udp_ip_port_mon.helper.h"
 #include "udp_ip_port_mon.parser.h"
 #include "udp_ip_port_mon.worker.h"
+#include "udp_ip_port_mon.icmp_mon.helper.h"
 //Main
 
 int main(int argc, char *argv[])
