+++ v1_Maintenance/lib/icmp_mon.helper.c
@@ -40,8 +40,8 @@ This file is part of MADCAT, the Mass Attack Detection Acceptance Tool.
 
 //Helper Functions
 
-#include "madcat.helper.c"
-#include "udp_ip_port_mon.icmp_mon.helper.c"
+#include "madcat.helper.h"
+#include "udp_ip_port_mon.icmp_mon.helper.h"
 
 void print_help_icmp(char* progname) //print help message
 {
