+++ v1_Maintenance/lib/udp_ip_port_mon.icmp_mon.helper.c
@@ -39,6 +39,7 @@ This file is part of MADCAT, the Mass Attack Detection Acceptance Tool.
 */
 
 #include "udp_ip_port_mon.icmp_mon.helper.h"
+#include "madcat.common.h"
 
 //saves and returns address of main buffer to be freed by signal handler
 void* saved_buffer(void * buffer)
@@ -59,4 +60,69 @@ void sig_handler(int signo)
     //exit parent process
     exit(signo);
     return;
+}
+
+void get_user_ids(struct user_t* user) //adapted example code from manpage getpwnam(3)
+{
+    struct passwd pwd;
+    struct passwd *result;
+    char *buf;
+    size_t bufsize;
+    int s;
+
+    bufsize = sysconf(_SC_GETPW_R_SIZE_MAX);
+    if (bufsize == -1)          // Value was indeterminate
+        bufsize = 16384;        // Should be more than enough
+
+    buf = CHECK(malloc(bufsize), != 0);
+    if (buf == NULL) {
+        perror("malloc");
+        exit(EXIT_FAILURE);
+    }
+
+    s = CHECK(getpwnam_r(user->name, &pwd, buf, bufsize, &result), == 0);
+
+    user->uid = pwd.pw_uid;
+    user->gid = pwd.pw_gid;
+    free(buf);
+    return;
+}
+
+char* json_do(bool init_or_reset, const char* format, ...)
+{
+
+    static json_struct json; //static to hold data in json_struct after return from function
+    static bool first_run = true;
+    signed int numchars = 0; //number of chars to write
+    va_list valst; //variable argument list
+    va_start (valst, format);
+    
+    if (init_or_reset) //should the json_struct be initialized or reseted?
+    {
+        if (!first_run)
+        {
+            free(json.str);
+            first_run = false;
+        }
+        CHECK(json.str = malloc(1), != 0);
+        *json.str = 0;  //add trailing \0 (empty string)                
+    }
+
+    //get number of chars to write
+    va_start (valst, format);
+    numchars = vsnprintf(NULL, 0, format, valst);
+    va_end(valst);
+   
+    //if an empty string has been provided as parameter, just return the pointer to actual string
+    if (numchars == 0) return json.str;
+
+    //allocate new memory for chars to write
+    CHECK(json.str = realloc(json.str, strlen(json.str) + numchars + 1), != 0);
+    
+    //append chars to string
+    va_start(valst, format);
+    CHECK(vsnprintf(json.str + strlen(json.str), numchars + 1 , format, valst), != 0);
+    va_end(valst);
+
+    return json.str; //return pointer to (new) string
 }
